refactor(skills): colocate remaining bundled helpers in skill directories
Move KTX, SRT, fine-dust, and setup helpers out of root bundle[] so CLI sync reads skill-local scripts/. Keep thin repo shims for PYTHONPATH=scripts tests.
